Production push

+++ app/services/v2/model_services/order_service.py
@@ -4,15 +4,17 @@ from models.v2.order import Order
 from services.v2.base_service import Base
 from models.v2.ItemInObject import ItemInObject
 from utils.globals import *
-from services.v2.database_service import DB, DatabaseService
+from services.v2.database_service import DatabaseService
+from services.v1 import data_provider
 
 
 class OrderService(Base):
-    def __init__(self, db: Type[DatabaseService] = None):
+    def __init__(self, db: Type[DatabaseService] = None, is_debug: bool = False):
+        self.is_debug = is_debug
         if db is not None:
             self.db = db
-        else:  # pragma: no cover
-            self.db = DB
+        else:
+            self.db = data_provider_v2.fetch_database()
         self.load()
 
     def get_all_orders(self) -> List[Order]:
@@ -48,21 +50,27 @@ class OrderService(Base):
                 return order
 
         with self.db.get_connection() as conn:
-            query = f"SELECT * FROM {Order.table_name()} WHERE id = {order_id}"
-            cursor = conn.execute(query)
-            order = cursor.fetchone()
-            if order:
+            query = f"SELECT * FROM {Order.table_name()} WHERE id = ?"
+            cursor = conn.execute(query, (order_id,))
+            order_row = cursor.fetchone()
+            if order_row:
+                column_names = [description[0] for description in cursor.description]
+                order = dict(zip(column_names, order_row))
                 query_items = f"SELECT item_id, amount, order_id FROM {order_items_table} WHERE order_id = {order_id}"
                 cursor = conn.execute(query_items)
                 all_order_items = cursor.fetchall()
-                order["items"] = all_order_items
+                # order["items"] = all_order_items
+                order["items"] = [
+                    ItemInObject(item_id=row[0], amount=row[1])
+                    for row in all_order_items
+                ]
                 return Order(**order)
         return None
 
     def get_items_in_order(self, order_id: int) -> List[ItemInObject]:
-        for x in self.data:
-            if x.id == order_id:
-                return x.items
+        for order in self.data:
+            if order.id == order_id:
+                return order.items
         return None
 
     def get_orders_in_shipment(self, shipment_id: int) -> List[Order]:
@@ -86,7 +94,7 @@ class OrderService(Base):
                 result.append(order)
         return result
 
-    def add_order(self, order: Order, closeConnection: bool = True) -> Order:
+    def add_order(self, order: Order, background_task=True) -> Order:
         if self.has_order_archived_entities(order):
             return None
 
@@ -122,14 +130,12 @@ class OrderService(Base):
                         items_insert_sql,
                         (order_id, order_items.item_id, order_items.amount),
                     )
-
-        # if closeConnection:
-        #     self.db.commit_and_close()
         self.data.append(order)
+        self.save(background_task)
         return order
 
     def update_order(
-        self, order_id: int, order: Order, closeConnection: bool = True
+        self, order_id: int, order: Order, background_task=True
     ) -> Order | None:
         if self.is_order_archived(
             order_id
@@ -139,7 +145,10 @@ class OrderService(Base):
             return None
 
         table_name = order.table_name()
-
+        order.id = order_id
+        if order.created_at is None:
+            existing_order = self.get_order(order_id)
+            order.created_at = existing_order.created_at
         order.updated_at = self.get_timestamp()
 
         fields = {}
@@ -148,17 +157,17 @@ class OrderService(Base):
                 fields[key] = value
 
         set_clause = ", ".join(f"{key} = ?" for key in fields if key != "id")
-        values = tuple(fields[key] for key in fields if key != "id") + (order.id,)
+        values = tuple(fields[key] for key in fields if key != "id") + (order_id,)
 
         update_sql = f"UPDATE {table_name} SET {set_clause} WHERE id = ?"
 
         with self.db.get_connection() as conn:
             conn.execute(update_sql, values)
-
-            if order.items:
+            existing_order = self.get_order(order_id)
+            if existing_order and existing_order.items != order.items:
                 delete_items_sql = f"""DELETE FROM {
                     order_items_table} WHERE order_id = ?"""
-                conn.execute(delete_items_sql, (order.id,))
+                conn.execute(delete_items_sql, (order_id,))
 
                 for order_items in order.items:
                     items_insert_sql = f"""
@@ -167,15 +176,13 @@ class OrderService(Base):
                     """
                     conn.execute(
                         items_insert_sql,
-                        (order.id, order_items.item_id, order_items.amount),
+                        (order_id, order_items.item_id, order_items.amount),
                     )
-
-        # if closeConnection:
-        #     self.db.commit_and_close()
 
         for i in range(len(self.data)):
             if self.data[i].id == order_id:
                 self.data[i] = order
+        self.save(background_task)
         return order
 
     def update_items_in_order(
@@ -194,6 +201,7 @@ class OrderService(Base):
     def update_orders_in_shipment(
         self, shipment_id: int, orders: List[Order]
     ) -> List[Order]:
+        updated_orders = []
         packed_orders = self.get_orders_in_shipment(shipment_id)
         for packed_order in packed_orders:
             if packed_order not in orders:
@@ -206,11 +214,10 @@ class OrderService(Base):
             if not order.is_archived:
                 order.shipment_id = shipment_id
                 order.order_status = "Packed"
-                self.update_order(order.id, order)
+                updated_orders.append(self.update_order(order.id, order))
+        return updated_orders
 
-        return orders
-
-    def archive_order(self, order_id: int, closeConnection: bool = True) -> bool:
+    def archive_order(self, order_id: int, background_task=True) -> Order | None:
         for i in range(len(self.data)):
             if self.data[i].id == order_id:
                 self.data[i].updated_at = self.get_timestamp()
@@ -231,13 +238,11 @@ class OrderService(Base):
 
                 with self.db.get_connection() as conn:
                     conn.execute(update_sql, values)
+                self.save(background_task)
+                return self.data[i]
+        return None
 
-                # if closeConnection:
-                #     self.db.commit_and_close()
-                return True
-        return False
-
-    def unarchive_order(self, order_id: int, closeConnection: bool = True) -> bool:
+    def unarchive_order(self, order_id: int, background_task=True) -> Order | None:
         for i in range(len(self.data)):
             if self.data[i].id == order_id:
                 self.data[i].updated_at = self.get_timestamp()
@@ -258,11 +263,22 @@ class OrderService(Base):
 
                 with self.db.get_connection() as conn:
                     conn.execute(update_sql, values)
+                self.save(background_task)
+                return self.data[i]
+        return None
 
-                # if closeConnection:
-                #     self.db.commit_and_close()
-                return True
-        return False
+    def save(self, background_task=True):
+        if not self.is_debug:
+
+            def call_v1_save_method():
+                data_provider.fetch_order_pool().save(
+                    [order.model_dump() for order in self.data]
+                )
+
+            if background_task:
+                data_provider_v2.fetch_background_tasks().add_task(call_v1_save_method)
+            else:
+                call_v1_save_method()
 
     def load(self):
         self.data = self.get_all_orders()
@@ -285,28 +301,24 @@ class OrderService(Base):
                         new_order.ship_to
                     )
                 )
-                print(has_archived_entities)
             if not has_archived_entities and new_order.bill_to is not None:
                 has_archived_entities = (
                     data_provider_v2.fetch_client_pool().is_client_archived(
                         new_order.bill_to
                     )
                 )
-                print(has_archived_entities)
             if not has_archived_entities and new_order.shipment_id is not None:
                 has_archived_entities = (
                     data_provider_v2.fetch_shipment_pool().is_shipment_archived(
                         new_order.shipment_id
                     )
                 )
-                print(has_archived_entities)
             if not has_archived_entities and new_order.warehouse_id is not None:
                 has_archived_entities = (
                     data_provider_v2.fetch_warehouse_pool().is_warehouse_archived(
                         new_order.warehouse_id
                     )
                 )
-                print(has_archived_entities)
         else:
             if new_order.ship_to != old_order.ship_to and new_order.ship_to is not None:
                 has_archived_entities = (
@@ -346,6 +358,21 @@ class OrderService(Base):
                 )
 
         return has_archived_entities
+
+    def check_if_order_transit(self, order_id: int) -> Order | None:
+        order = self.get_order(order_id)
+        shipments = data_provider_v2.fetch_shipment_pool().get_shipments_for_order(
+            order_id
+        )
+        can_change_to_Transit = True
+        for shipment in shipments:
+            if shipment.shipment_status != "Transit":
+                can_change_to_Transit = False
+                break
+        if can_change_to_Transit:
+            order.order_status = "Shipped"
+            return self.update_order(order_id, order)
+        return None
 
     def check_if_order_delivered(self, order_id: int) -> Order | None:
         order = self.get_order(order_id)
